Resolve the issue of Secured annotation and webflow
We knew that the @Secured annotation no longer supports being defined on controller action closures since Spring Security 2.x [1].
So we will have to convert them to real methods. That's fine in the case of normal methods.

Besides, we are employing webflow plugin that supports the creation of web flows built on the Spring Web Flow. To create a flow,
we should add an action that ends with the convention Flow. Currently, we define web flows as closures instead of methods.
According to changes in Spring Security 2.x that particulary addresses to @Secured annotation, these web flows should be converted
to methods as well. But somehow they cannot work.

To solve this problem, we've come back the previous state keeping web flows as closures and setting @Secured annotation at the top of the class.
This is enable to define web flows as closures and allows us to customise the security level on each of the methods. For example, we want
download() method to be visible for everyone, so it should be secured by using IS_AUTHENTICATED_ANONYMOUS annotation at the top of method.

References:
[1] https://grails-plugins.github.io/grails-spring-security-core/v2/guide/newInV2.html