Forced flushing of entries accessed from other transactions.
Explicitly flush changes to objects that must be accessed in another
transaction. Until flush is performed, said changes remain present only
in the Hibernate session associated with the current transaction, and
thus are not visible in other transactions [because every tx has its own
hibernate session and associated persistence context].

Also commented out calls to publish events related to model or revision
creation from the transactions they are created in. This is because the
new model/revision is indexed in a separate transaction. Also,
creating the transport command objects for the new model or revision
triggered a call to modelHistoryService.addToHistory, which meant that
new transactions are not able to access the domain objects.