Prevented concurrent AclSid creation during model submission.
ModelService methods to upload or update a model also need to add the
appropriate ACLs for the newly-created domain objects. If concurrent
requests attempt to insert two AclSid objects with the same principal
Spring Security does not complain, but the database will because of the
uniqueness constraint.

Therefore, guard ACL creation with a Reentrant lock.