Disable secondary dependency resolution for plugins.
This is an artefact from the time when Apache Ivy was used for
dependency resolution and Grails still allowed dependencies to be added
via the install-plugin command.

Since we now use Aether for dependency resolution, this doubles the time
it takes to configure the project classpath without adding any benefits.