JBM-36 first stab at showing ModelOfTheMonth info.
Entries that are the object of BioModels' ModelOfTheMonth feature are
described in a manner that aids comprehension for non-specialist
readers.

Added machinery for capturing that models are referenced in a MoM entry.

In order to display this information we need to invoke the newly-created
tag library from the relevant view.