JBM-43 clear Hibernate session before importing revision.
ModelService.addValidatedRevision expects a clean persistence context,
but the batch importer uses a shared session for processing all models.

This meant that when modelService attempted to fetch the revision object
from the database and attach it to the current persistence context it
failed due to the session already containing references to the model
object as it looked after it was first submitted.