fix for JBM-129 show the correct curation status.
The previous approach traversed through the full set of annotations
in order to find the matching one. This doesn't scale well for models
that are rich in annotations.

A better approach is to run a database query which joins the annotation
tables, since it would use indices.

For now, rely on the presence of the publication identifier as a marker.