Replaced DomainAdapter facade.
There was an assumption that calls to
    DomainAdapter.getAdapter().toCommandObject()

would not incur a major performance hit since we relied on Groovy's
caching of previous calls to invokeMethod(). This does not appear to be
the case, so this commit sacrifices the convenience of consistency over
speed.

We replaced calls to figure out the adapter and invoke toCommandObject()
via dynamic invocation with instantiation of the appropriate adapter,
meaning that not only do we benefit from the speed improvements, but
this also paves the way for enforcing static compilation of the bits of
code where we instantiate command objects.

Also improved the logic of creating ModelTransportCommands for admins by
bypassing calls to modelService.getLatestRevision. This is useful in the
context of regenerating the search index, as that process triggers such
a call for every revision (i.e. O(n^2) complexity).