JBM-186 simplified regex for model identifiers in URL mapping.
The regexes for every model identifier scheme were OR'ed and wrapped
around brackets. The latter simply implies grouping, but we don't
reference this group in the regex, so it is unnecessary. This commit
removes the extraneous brackets, which saves us from needing string
interpolation to figure out the regex for model identifiers.