JBM-179 use ajax to save data without reloading page
Because simulation results have curation image and the other textual/date data,
it is a challenge to update them at the same time in one click. We split the task
into individual tasks. On the one hand, the curation image should be updated
immediately once uploading a new image. On the other hand, the other
fields such as comment, submitter, last modifier, etc. are sent to the database
if curator hits on Save button.

TODO:
- implement autocompletion for submitter and last modifier boxes.
- use datetime picker for datetime-related fields.