JBM-80 take the main files in working if available
We introduced a map to retain the latest updated main files during upload
flow before transferring it to ModelController. The main files being
displayed in the uploadFiles page will be taken from this map first if it
is available, otherwise they will populate in the main_files variable as
initialised from the existing revision at the initial stage of
updating process.