Turned CacheService and ModelClassifierService into non-transactional beans.
Disable transactional behaviour as that leads to nested transactional contexts
(i.e. the outer tx gets paused while the inner one is in progress).

Also switch back from static methods to instance methods for these services
to avoid AOP issues.

Make CacheService an application-scoped bean because it is expected to exist for
the duration of the application, whereas singleton-scoped bean instances can be
garbage collected and reinstantiated during the application's lifecycle.

Co-authored-by: Tu Vu <glmanhtu@gmail.com>