JBM-179 resolve the bug on refreshing the curation notes editor page
After updating and adding the curation notes, we sometimes hit Fresh
button to see what has been updated/persisted. The issue could happen
is not to show the curation notes that has been added because the url
does not consist of curation note identifier in case we add a new
record. In the mean time, we would keep showing the messase telling
the status of the operation which is successful or failed.

To solve this issue, we hack to update the url by adding curation
notes identifier as the url's last param without reloading the page.