JBM-266 deal with different model types with the unique filter
We are considering three types of model which are private, shared and public.
Because using only one filter variable to capture the facet which has been
clicked, we have to differentiate the model type with the other facets. The
model format and curation status, for instance, can use 'like' operator
while the type of model needs a more sophisticated query.