JBM-285 introduce a new attribute to additional files map
To track of which the additional input file is processing, we need to add
a new attribute into the map. The new attribute maintains the identifier of
the additional input file that is helping us to know the element in process. Thus,
we used the identifier to check the state of every input file once an event
fires on it, its associated description and Discard button.

The patch is a part of JBM-80