JBM-285 take the advantages of multipart file upload functionality
Passing a JSON string from the client-side to the server side by storing it
into a temporarily hidden field seems to be unstable and is unable to cope with
the file description containing single or double quotes. By declaring
input file and input text description with the same name, we are easy to
transfer data from client to server via params and request object.

The solutions allows to release the maintenance of the map of additional
files as well as populating JSON string as stated above.

This work has contributed to complete the ticket JBM-80.