JBM-266 query performance improvement for showing My Models.
The query in modelService.getAllModels() was using full table scans
and was joining model, revision, format, user and acl_* in a way
that caused temporary tables to be created.

This commit reworks the queries to reduce the number of joins and to
perform lookups based on entity identifiers, rather than any other
attribute.

Also made the code slightly easier to read by removing some of the nested
if-else constructs and simplifying some of the logic.