JBM-285 swap two scenarios of if statement to fix the bug
To cope with the removal and replacement of the main file, we have to
check the main file which is stored in UploadFilesCommand first. The
reason is that the main file is a required field which the file upload
input is always initialised regardless of either uploading a new file or
storing the uploaded file to the working memory. The checking this field
ensures that the newly uploaded file is always taken into account first.

This fix supports for JBM-80.