JBM-291 bring the latest updates to the new BioModels platform
By applying model change detectors, that helps us to determine which
models should be taken from the old system and added a new revision to
the corresponding model existing in the new system.

The patch has not yet coped with the latest modifier instead the
administrator as the submitter as well as the actual committed message
provided by the  latest modifier. In additiona, it has not yet dealt
with the additional files which should be brought alongside all revisions.

This work contributes to the solution to the ticket JBM-244.