JBM-88 find a way to add the additional files to the subsequent revisions
The additional files are the ones which were submitted by the model owner in
the initial submission. They are also the files created by curators during
the curation pipeline. To make them along the subsequent revisions, we did
* make the snippet for finding the additional files submitted by the
  submitter or COPASI/SED-ML file used in the curation process available
  for reusage it in other places.
* recall this snippet when adding the latest revision.

This work contributes to the completion of the tickets JBM-200, JBM-244
and JBM-291.