JBM-291 relaxed logic for mapping users from the internal comments.
We parse the internal comments for all processed models, but the
author & changelog info is only needed for the most recent ones, those
that have been updated in the old system since we last imported them in
JUMMP (this is why the user mapping for internal comments has only a few
entries).

Previously if we failed to extract any information from the comments for
a model parseCurationCommentsForModel would defensively return null.
This commit relaxes the logic so that parseCurationCommentsForModel
returns the information it was able to gather. This works fine because
we already have sensible defaults.