JBM-299 remove non-public endpoint from API docs.
The 'My Models' page is only available to users that are logged in.
Until the API handles programmatic authentication (e.g. using OAuth)
there is no point in exposing this endpoint from the documentation.

Granted, one could programmatically log in, store the session cookie
and include it in the /models request, but this is a hack rather than
the normal behaviour.