JBM-299 allow content negotiation for Firefox&Chrome.
This is based on the Accept request header. According to the Grails 2.5.5
docs[1] on content negotiation, the value of this header is ignored for
requests made using any of the four main browsers due to the fact that
historically they were using the header Accept */*, which is not very
helpful.

Modern browsers have updated their content negotiation strategy, so it's
okay to rely on the Accept header information they provide.

[1] https://grails.github.io/grails2-doc/2.5.5/guide/single.html#contentNegotiation