JBM-299 remove old swagger integration.
The old approach attempted to automatically generate the API docs from
controller annotations and url mappings, but it included a lot of Groovy
cruft and it incorrectly interpreted the /modelId endpoint as requiring an
'action' parameter, which it the ignored.

We now rely on maintaining the swagger json file + static HTML pages
in web-app/docs.