JBM-308 refactored method for testing if the current user is a curator.
Move it away from ModelController to UserService because logic should
reside in services rather than controllers.

Also simplified the way the test is performed: rather than iterating
over all the roles assigned to a user (linear process), we check if
there is a corresponding UserRole entry, which is constant time
operation.