JBM-298,JBM-270,JBM-311 fix message when publishing non-curated models.
The notification shown after the model is published should contain the
publication identifier, if one was generated.

Before this commit we would check if the revision is public to decide if
the publicationId should be shown; this meant that we'd try to show it
for all successfully published revisions, rather than only those that
are curated and published.

Moreover, we would show the publicationId by relying on RevisionTC's
modelIdentifier() method, which falls back on the submission id if no
publication identifier is present.

This commit ensures that we only show the publication identifier if it
is defined, which normally entails that the model is curated and
published and that the model identifier settings include a section for
publication identifiers.