Remove SbmlCache type parameterisation.
The old implementation was using generics, but the semantics were only
working as expected (or close enough anyway) if the values were Longs
and the keys were RevisionTransportCommands (RTCs).

RTCs are inherently transient, so internally, the cache was using the
revision id to deal with the fact that the same revision can have two
RTC representations that are not identical (because they will point to
different file checkouts).

This abstraction leaked in the entrySet()/keySet() methods because we
were trying to create an RTC for the revision id but with all other
fields undefined.

For these reasons, this commit removed the parametrisation for SbmlCache,
although generics are still used internally. Since RTC has lots of fields
of complex types (Objects, rather than primitives), there is significant
space overhead when used as a key when compared to a Long. Therefore I
make the SbmlCache keys the revision IDs.