JBM-276 overhaul management of model id generator beans.
Before this commit, id generator settings were parsed in Config.groovy
and then subsequently turned into bean definitions in resources.groovy.
The parsed settings where stored in Config with the key jummp.model.id,
and we removed that setting in resources.groovy as soon as we finished
creating the (singleton) generator beans (e.g. submissionIdGenerator,
publicationIdGenerator).

The generator beans were initialised once, at the start of the application,
with the latest identifiers in the database and they would update this
value every time their generate() method was invoked.

This can cause issues if a bean is not used for a long time, as Spring
does not preclude the Java garbage collector from destroying it. The
next time that bean is needed Spring simply recreates it using the bean
defintion.

The problem is not only that we 'defensively' deleted the parsed id
generator settings, meaning the last known id created by a generator
bean was ignored, it's also very likely that model identifiers have
been created since we last looked in the database, so the original
seed value is now incorrect.

This commit changes the way id generator beans get defined, initialised
and managed by Spring:

Model ID generators
-------------------
Generator beans now have prototype scope, so every time we request one
from the application context we get given a new instace. The beans are
now defined with type ModelIdentifierGeneratorFactoryBean. When Spring
sees FactoryBean definitions it uses the logic defined by the factory
to instantiate the bean. This helps us ensure that the generators we
get from the application context parse the model identifier settings
defined in JUMMP's property file. See section 5.8.3 of the Spring
docs[1].

ModelIdentifierGeneratorInitializers are prototype-scoped beans which
extract the most recent value that a given id generator has produced.
The id generator factory bean uses this information when it constructs
the generator instance.

Care is needed when introducing a prototype-scoped dependency into a
singleton bean. Spring pre-instantiates and caches singleton instances,
meaning that the singleton will always reuse the same instance of the
dependency. For this reason, modelService and modelDelegateService no
longer declare the id generators as dependencies. Instead, modelService
provides getSubmissionIdGenerator() and getPublicationIdGenerator()
methods which request eponym beans (e.g. submissionIdGenerator) from
the applicationContext. An alternative to this would be to make the id
generator factory bean an ObjectFactoryCreatingFactoryBean[2].

Model ID decorators
-------------------
ID decorators now no longer use the application context to publish
changes to their values. Instead, they have an explicit reference to
the generator that they work for and they notify the generator explicitly.

Optimisations
-------------
To speed things up and tackle the overhead of creating id generators on
the fly, I enabled static compilation for decorators and generators.

[1] https://docs.spring.io/spring/docs/3.2.x/spring-framework-reference/html/beans.html
[2] https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/beans/factory/config/ObjectFactoryCreatingFactoryBean.html