JBM-282 fix tests of ModelClassifierService by chance
When working on Model of The Month, we accidentally found the unit tests
of ModelClassifierService do not work. @MockFor(ModelClassifierService)
does not initialise the modelClassifierService bean properly and its
afterPropertiesSet() method does not seem to be called. We therefore
manually define the bean using doWithSpring and instantiate it in setup().

Co-authored-by: Mihai Glont <mihai.glont@gmail.com>