JBM-311 ensure that we always index the newly created revision
By adding a new publication identifier to the current revision, we have to
create a new revision where the new identifier has been added to as an
annotation. We also make sure that the indexing service is launched against
this revision in order to avoid missing annotations from the previous revision.

Calling indexing service in doBeforePublishingCuratedRevision can cope with
two situations:
* The model in operation is private and has been curated. We are now publishing
  it.
* The model in operation is public and non-curated. We are now adding curation
  notes and then changing its curation status.