JBM-311 populate the newly-created publication identifier
This work aims to populate the publication identifier and send to the view
via the response object. In the view, we use this identifier to create the
redirected link to the model page. The page is accessed by the publication
identifier.

We need to keep users staying on the current page for a few seconds to
see the message when updating of the curation status before redirecting to
the web page accessing via the publication identifier.