JBM-262 fix compiling  and running errors
JummpHttpService could not be resolved during the compilcation if it is
stored under grails-app/utils. To resolve the issue temporarily, it is
moved to utils directory of core-api plugin.

In addition, this patch fixed the issue of non-static variable referencing
to static method. The issues emerge at starting the application.