JBM-345 use Java built-in libary to escape XML special characters
Replacing CDATA section with StringEscapeUntils library, the title
and description of Model of The Month's entries are rendered properly.

According to RSS Board's website, the link and url attribute with
non-ASCII characters (e.g. '&') should be percent-encoded. On the one hand,
encoding '&' characters leads to the coruption in retrieving
these links with some RSS aggregators such as Firefox, Thunderbird.
On the other hand, we couldn't please all RSS readers. Thus, we decided
to leave the link with the '&' symbol unchanged as what it is.