JBM-363 cope with model container overflow.
JUMMP is capable of switching to a new container if the current one is
full, but until this commit it assumed that a container would never have
more entries than the allowed threshold (1000 entries).

This commit addresses this limitation by relaxing the condition for
triggering a model container switch from exactly 1000 to at least 1000
entries in the present one. When an overflow is detected we also now log
an error message.

Fixes JBM-363