JBM-364 reduce risks of duplicating publication without link
We had just looked up publication in the database when the link provider
is different from manual entry type. As a result, an empty form is
always rendered if the submitter has selected 'Publication without
link'. Why wouldn't we have shown the persisted publication regardless
of its types?

What we have improved are:

* to verify the way of detecting the fact that user has changed
  the type of the publication.
* to reload the publication without link of the model in question if the
  model was added the publication from the last updates.
* to ensure that users could change the other types of publication (see.
 JBM-295).