JBM-364 improve the way of finding publication
Searching records in publication table before having a decision of
updating or creating a new entry is vital in management of publications.
Based on a given transport command object, we will know its status that
is holding a persisted publication or is a completely new entry. Thanks
to the id property introduced from the commit 8354565bc and the
improvements committed in 4ec6e44be, publications are now searched by
this id first, otherwise the search will be relied on the link as
before.

This patch also allows us to prevent from creating a new publication
once submitters add the publication manually.