JBM-364 fix the issue 'Timeout trying to lock table'
This issue was caused by persisting publication authors when the service
fetches the publication metadata from PubMed Central. The changes
introduced in this commit aim at performing the following purposes:

* Don't save person authors into the database at parsing a JSON string
containing the information of the publication authors. To do so, we
could eliminate calls of PersonAdapter to convert domain to transport
command objects. Also, the working flow in the insertion and update of
publication is consistent in which we don't try to persist data in
transitional steps.

* Remove Transactional annotation set for fetchPublicationData method.

* Come back to `grails.transaction.Transactional`.

* Refactor `reconcile` method and order it closer the other private
methods.

* Better `fromCommandObject` method.

This patch also aims to revert the commit 11dd8b5.

Signed-off by: Mihai Glonț <mglont@ebi.ac.uk>