JBM-364 throw and catch exceptions happening with bad authors
When editing Publication details, users might enter invalid authors.
This patch fixed the bugs raising if a few of authors have no name or an
invalid orcid which is not matched to the defined pattern.

We introduced new exception classes to handle invalid data. The
exception is thrown from server, and then caught at client. Also, the
friendly error messages will be shown at the client once there is an
error.