JBM-364 ensure that the publication link is changed
Since the commit 21c8b20, looking for publications is mainly performed
via the record identifier if the transport command object has an identifier.
However, this won't work with the updating model publication by changing
the publication link. We turn the logic in the other way round to ensure
that the latest publication is fetched.