JBM-382 check publication details of model when requesting to publish
This commit partially implements the requirement addressed in the story
JBM-381. That is to check the existence of the publication details when
receiving a request for publishing a model. If the model has a publication
reference, curators will publish the model. Otherwise, we argue users to
update their model and add minimal metadata of the publication reference.