BP-68 simplified downloading of parameter search results.
Fixes bugs that break the feature in browsers that are not Chrome-based.

Also avoid the extra roundtrip required by the AJAX callback of the
previous approach by simply forwarding the server response to the
browser.

I've also added a bit more error handling around the call to download
files so that the download button is re-enabled even if there is an
error with one download request.