JBM-387 use grails data binding for Tag and ModelTag
Instead of populating and rendering a set of tag names, a set of
Transport Command objects (i.e., TagTransportCommand and
ModelTagTransportCommand) to be built in MetadataDelegateService is
transferred to model display page via ModelController.show. After
changing and updating this set of TagTransportCommand objects in the
model display page, clicking on Save button will send an updated version
of this set with the model in question to ModelTagController so as to
leverage Grails data binding.

This change causes the modifications to the views of showing persisted
tags for the given model and initialising select tag element.

This patch also includes improvements of saving model tag associations
in batch mode. We use the mechanism of deleting the existing tags then
inserting all the newly updated ones in place of finding the delta
between them.

At the time, we disable adding new tag dynamically. Once this feature
is enabled, we need to silently reload the data source of select2 by either
triggering another ajax call to fetch newly inserted tags or changing
the way of initialising select2's data via ajax [1] rather array as we
have done.

This work is the implementation to JBM-390.

Refs:
[1] https://select2.org/data-sources/ajax