JBM-406 make sure identifiers of persisted authors populated
To reconcile the authors of a publication during update flow, we are
trying to determine authors not being in the list of to-be-added authors
via their identifier or orcid. This principle works fine unless the
authors have no orcid. Therefore, we must bring the id of persons
associated with the authors along the other information so that authors
can be reused. Moreover, grails databinding won't cause errors.

This commit also contributes to fix JBM-407