Use prototype scope for ModelConversionService.
We need to store state (export folder location, conversion service URL)
inside ModelConversionService. A spike in memory usage would trigger a
garbage collection which would reclaim infrequently-used objects. If the
(singleton) modelConversionService bean instance is retrieved after a
GC, the new instance then loses its state.

To fix this, we change from singleton to prototype scope, meaning that
every time the bean is requested we get a new instance. The service now
implements GrailsConfigurationAware and we use setConfiguration(),
which is called during the construction of the bean instance, to
initialise the bean's state appropriately.