only include tag identifier into dynamic finder if it is available
When assigning tags for models in Path2Models via batch importer
(ref. JBM-400 and JBM-416), we found this defeat causing of creating
a new tag all the time without reusing existing ones. This commit
checked the transport command object of Tag is new or found from the
database. Don't include identifier into the dynamic finder without
checking whether it is null or not. See works for JBM-387.