JBM-393 bump logging to aid troubleshooting.
In some circumstances trying to create RevisionTransportCommands for
modelDelegateService.getAllRevisions() results in a
LazyInitializationException (LIE).

Since we're caching modelService.findByPerennialId() results (default
TTL is 2 minutes), when there is a cache hit the output of
modelService.getAllRevisions() will return Revisions which are still
attached to the Hibernate session from when they were added to the cache
-- in effect they are detached from the current session.

These detached Revisions' format association is an unresolved proxy due
to the fact that Grails by default does not unwrap one-to-many
associations (they are lazily loaded). This is fine so long as we access
those associations in the same transaction, but if the session is
closed, then a LIE is thrown.

The thing is I cannot reproduce the LIE locally, so this commit includes
additional logging to aid debugging.