Avoid inadvertent config overwrite in plugin's doWithSpring block.
Plugins' doWithSpring closures get invoked after the main app's
Config.groovy, but before the latter's resources.groovy.

This caused the BioModels DOM plugin to accidentally overwrite a setting
defined in the main application's Config.groovy as opposed to amending
it.

This commit ensures that the plugin in question only appends to the main
configuration if necessary.