JBM-429 relax search criteria for GORM dynamic finder
When working on JBM-429, we found that linkType is always unique.
Moreover, we have never had a chance of providing pattern as well as
identifiers prefix of the publication link provider directly such that
GORM dynamic finder based on three properties can find a value from the
database. In other words, this dynamic finder always returns null.
Therefore, it is unnecessary to find a link provider on three different fields.

AnnotationStore should be updated on linkType of PublicationLinkProvider