make fetching model level annotations more robust
In some use cases, we don't need to pass a whole Revision object instead of its identifier.
Why do we need to convert Revision to command object, then pass it to the method.
Such a behaviour is expensive and error-prone.

This improvement also helps us preventing the error of bad SQL grammar which is caused
by logic used in RevisionAdapter.toCommandObject()