JBM-348 : Added a link which will open reactome pathway in a modal or won't if no reactomeId is associated with model
- Added a link which will open a modal containing the Pathway diagram associated
with respected model id
- If no association found then it won't open the modal box
- There is one pending issue in this which is when autoOpen is set to false
then diagramjs library of reactome gives error and does not display the pathway diagram.
I tried manually closing the modal box and keeping autoOpen as default which is 'true'
but it did not work, will need to look into this. Now the popup will open as user opens this page.