JBM-447 made decorationService.getRecentlyAccessedModels() leaner.
The query used before was identified as a potential performance bottleneck.

In response, this patch introduces the following changes:

1. The method no longer creates ModelTransportCommand instances
This was causing full table scans in the model table and the call to
create MTCs is expensive.

2. We no longer compute model hit counts
They were not shown to the user, yet they account for ~40% of query time.

3. We ensure all query params are initialised correctly
Previously, "then" could be uninitialised.

4. We replace tabs with spaces for consistency