JBM-437 do not delete the files of the previous revision.
During model updates which involve adding new files and removing exiting ones,
GitManager currently first performs the addition and only then the removal.
This means that if a file appears in both sets it will be deleted altogether,
and its new version will therefore be lost. This explains one of the issues
in JBM-472.

The fix is to simply not include the repository files of the previous revision
in the set of files that should be deleted when we update the model using the
Uhlen updater script.