JBM-479 revert to extension detection for Mathematica.
The suggested fix from https://issues.apache.org/jira/browse/TIKA-3034
does not work, in the sense that Grails does not seem to allow us to
customise the classpath, even though there is a SetClasspath event which
I tried to intercept using _Events. No matter what I tried, the custom
MIME type declarations file was not being added to the project classpath
in front of the tika-core JAR, causing Tika to therefore ignore it.

We should revisit this in the future, once a new version of Tika is out.