JBM-479 fix submission warning for Matlab submissions.
We forgot to register matlab as a format for which we ignore the version
for the purposes of validation during the submission process.

Added some code in CommonFormat for simplifying the logic in
ModelFormatAdapter, but it causes circular dependencies between core-api
and the common-format plugin. I left a TODO for this in
ModelFormatAdapter.