JBM-451 give a chance to the service in order to rebuild model cache
Recently, there have been some reports about missing files. Although the revision
cache directory exists, there is no files in it. The former implementation only throws
an instance of ModelException as long as the revision cache directory does not exist.
This improvement has allowed the callers to throw a ModelException, even though no
files of the revision were copied to the cache directory.

In addition, to log informative messages is necessary to debug the issue in the future.