JBM-347 try to use rest-client-builder to consume BioModels Restful API
This work aims at building statistical data for plots shown on the front page. The key idea
is to fetch data about the aspects mentioned in forthcoming lines from searching them in
BioModels. It means that using queries to search data from BioModels ensures inconsistent
information displayed on the plots and search results.

The data could be built replied on this route are curation states, modelling approaches, organisms.
Also, we need the total of submissions, total of models in each domains, total of classes shown
in GO chart and total or records in Parameter Search that everything will be cached on Redis Server
and loaded from it to display in the front page.

There have been several approaches of consuming web services but we are choosing a simpler way
to implement rapidly and a lightweight library to be integrated in jummp.