do a complete refactoring to the publication view
It does not need to check model.publication again because
it has done in the parent view, i.e. modelDisplay before rendering
the showPublication template view.

Shorten model.publication expression by introducing a local variable

Separate show and edit view to ease future extensions and maintainability

Show a link to access the show view of the publication which
the link type is not PUBMED and DOI. Such a publication is manually
captured in BioModels. Showing users a presentation of that publication
is better than linking to the model itself.