fix the menu of My Account hidden in medium and large screens
Foundation dropdown menu adds classes opens-left, opens-inner and opens-right
automatically using JavaScript. Inspecting the manipulation, it is noticed that the dropdown
menu always is opened inner, therefore, it is partially overflowed to the right on small, medium
and large screens. This could be a bug as discussed on [1].

This provisional solution just tries to replace opens-inner with opens-left after the mouse over event
immediately occurred.

Refs
[1] https://github.com/foundation/foundation-sites/issues/7251