JBM-347 add an action to update Redis Cache on a single run
Apparently, there are separate actions for every update coming alongside widgets.
For example:

* The updateDataNewsWidget action will run the service
to update the data of the News widget.
* The update updateMoMEntryOnRedisCache action will invoke
the service to update the data of the MoM widget.

However, we haven't implemented yet any action to update data of
all widgets on the home page on a single invocation. This action
might be handy in some use cases. For example, this allows us to
run the curl command to update Redis Cache which is appended to the
deployment script to be called after the deployment succeeded.