JBM-511 make use of the published date instead of the created date
The created date is when the item was created and is used for management.
When the item is published and publicly shown at the front, the published date
(i.e. publishFrom) makes much sense of sorting and displaying. By and large,
audiences just care about the date when the item was published.