JBM-487 synchronise the last used value at getting the bean
The getSubmissionIdGenerator or getPublicationIdGenerator method will
hit ModelIdentifierGeneratorFactoryBean to obtain the bean where it is
initialised the next value based on the last used value to be retrieved either
from the database or the Redis Cache. This last used value will decide which
is the next value to be used in the Decorator classese. Therefore, we have to
synchronise the last used value to ensure that multiple threads see the same
updated value.