JBM-375 refactor Dockerfile.
Up to now, we were building the Docker image from Docker, which meant that
project dependencies would be downloaded after any code change or run-app
execution.

This commit introduces a build script written in BASH which produces the
war file on the host machine and deploys it on a Tomcat 7 container.
Sensitive build arguments can be passed as command line arguments when
invoking the script.