JBM-527 process the post submission
This commit resolves several unavailable services which are happened
after a successful submission. These services include of sending an email
to the submitter, sending a notification to the model collaborators, displaying
the appropriate messages at the end of the flow...

TODO: figure out the error when using Promises.task to invoke
processPostSubmission service.