JBM-519 reconcile and enhance the selection of publication source
This refactoring wants to redesign the section of selecting a publication source
which is done via either the step of adding publication record during submission or
update flow as well as the action of adding and editing a publication record.

The section comes along with the publication form. The complexity is to handle
various use cases of changing publication sources.